requirements.txt
diffusers 
transformers 
scipy 
ftfy
flax==0.5.0
ipywidgets 
msgpack 
rich
Flask==2.0.2
streamlit==1.5.0
Pillow==8.4.0
